+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - WFS-T: Multi-update works correctly when multiple layers are configured in the WFS-T tool.
 - i18next: Updated local storage key for i18n to include a portal identifier, ensuring unique language settings for different portals on the same server.
 - SearchInterfaceSpecialWfs: Improved namespace handling and coordinate transformation. Fixed issues with WFS servers using different namespace prefixes and EPSG:4326 point coordinates.
+- UrlParams: Fixed zIndex of second baselayer, when share link with mdid and multiple baselayers is opened.
 
 ---